Support for libgpkg for vector data in Geopackage

+++ SpatialConnect/SCServiceStatusEvent.m
@@ -39,6 +39,6 @@
 }
 
 - (NSString *)description {
-  return [NSString stringWithFormat:@"%ld,%@", self.status, self.serviceName];
+  return [NSString stringWithFormat:@"%ld,%@", (long)self.status, self.serviceName];
 }
 @end
